refactor(server): collapse refresh-token data-access seam
Fold the SQL primitives from db/refresh-token-queries.ts into
repositories/refresh-token-repository.ts as private internals, producing
one deep module whose public surface deals exclusively in raw tokens.
SHA-256 hashing, expiry parsing, and the rotation transaction all live
behind the seam; rotateRefreshTokenTx is no longer exported and no longer
takes a pre-hashed token — hashing policy no longer leaks into the SQL
layer.

The public surface (generateRefreshToken, validateRefreshToken,
revokeRefreshToken, rotateRefreshToken, revokeAllUserTokens,
cleanupExpiredTokens, parseRefreshTokenExpiry) is unchanged, so
routes/auth.ts needs no edits. RefreshTokenRow is now an internal type.

db/refresh-token-queries.ts and its test are removed; the SQL-level
assertions are migrated into refresh-token-repository.test.ts, which now
mocks the DB (query/transaction) directly.

Refs #1235 (C2, prefactor for the Session module — #1237)
